+++ src/main/java/org/sopt/dto/ErrorResponse.java
@@ -1,4 +1,4 @@
-package org.sopt.global;
+package org.sopt.dto;
 
 public record ErrorResponse(int status, String message) {
 }
