+++ src/main/java/org/sopt/service/PostService.java
@@ -4,8 +4,8 @@ package org.sopt.service;
 import org.sopt.domain.Post;
 import org.sopt.domain.User;
 import org.sopt.dto.*;
-import org.sopt.global.CustomException;
-import org.sopt.global.ErrorCode;
+import org.sopt.global.exception.CustomException;
+import org.sopt.global.exception.ErrorCode;
 import org.sopt.repository.PostRepository;
 import org.sopt.repository.UserRepository;
 import org.springframework.stereotype.Service;
